카카오 서버에 연결 해제 요청 추가

+++ project/user/views.py
@@ -343,6 +343,12 @@ class KakaoConnectView(APIView):
     @swagger_auto_schema(
         operation_description="카카오 계정 연결 해제하기.",
         manual_parameters=[jwt_header],
+        request_body=openapi.Schema(
+            type=openapi.TYPE_OBJECT,
+            properties={
+                "access_token": openapi.Schema(type=openapi.TYPE_STRING),
+            },
+        ),
     )
     def delete(self, request):
         if not hasattr(request.user, "kakao"):
@@ -350,7 +356,27 @@ class KakaoConnectView(APIView):
                 status=status.HTTP_400_BAD_REQUEST, data="연결된 카카오 계정이 없습니다."
             )
         kakao = request.user.kakao
+
+        access_token = request.data.get("access_token")
+        if not access_token:
+            return Response(
+                status=status.HTTP_400_BAD_REQUEST, data="access_token을 입력해주세요."
+            )
+
+        response = requests.post(
+            "https://kapi.kakao.com/v1/user/unlink",
+            headers={"Authorization": f"Bearer ${access_token}"},
+            data={"target_id_type": "user_id", "target_id": kakao.identifier},
+        )
+
+        # 요청 처리되었는지 확인
+        if not response.data.get("id") == kakao.identifier:
+            return Response(
+                status=status.HTTP_400_BAD_REQUEST, data="access_token이 유효하지 않습니다."
+            )
+
         kakao.delete()
+
         return Response(status=status.HTTP_204_NO_CONTENT)
